Agent Runner: The task from the last commit was not done fully, can y... (#118)
* [REFACTOR] Task 024: Create theme directory structure and base files

- Created agents_runner/ui/themes/ directory
- Added subdirectories for codex, claude, gemini, and copilot themes
- Created __init__.py files with appropriate docstrings
- Structure follows Python package conventions
- No breaking changes to existing code

* [REFACTOR] Extract Gemini theme into dedicated module

- Create agents_runner/ui/themes/gemini/background.py
- Move _GeminiChromaOrb dataclass to new module
- Convert Gemini methods to module-level functions:
  - gemini_palette()
  - ensure_gemini_orbs()
  - constrain_gemini_orbs()
  - tick_gemini_chroma_orbs()
  - paint_gemini_background()
- Update graphics.py to delegate to new module
- All functions accept necessary state as parameters
- No functionality changes (visual output identical)
- Reduces graphics.py complexity and improves modularity

Task: task-025-extract-gemini-theme

* [REFACTOR] Extract Claude theme implementation into dedicated module

Moved Claude theme implementation from graphics.py to agents_runner/ui/themes/claude/background.py.

Changes:
- Created agents_runner/ui/themes/claude/background.py with all Claude-specific code
- Moved dataclasses: _ClaudeBranchTip, _ClaudeBranchSegment
- Moved constants: _CLAUDE_SEGMENT_LIFETIME_S, _CLAUDE_SEGMENT_FADE_IN_S
- Converted methods to module-level functions:
  - _claude_palette() -> claude_palette()
  - _ensure_claude_tree() -> ensure_claude_tree()
  - _reset_claude_tree() -> reset_claude_tree()
  - _tick_claude_tree() -> tick_claude_tree()
  - _paint_claude_background() -> paint_claude_background()
- Updated graphics.py to import and delegate to claude.background module
- Removed 212 lines from graphics.py (1407 -> 1195 lines)

Follows pattern established by Gemini theme extraction (task 025).
No functionality changes - visual output remains identical.

* [REFACTOR] Extract Codex theme to dedicated module

Extract Codex theme implementation from graphics.py into
agents_runner/ui/themes/codex/background.py module.

Changes:
- Create codex/background.py with all Codex-specific rendering code
- Move blend_colors, get_top_band_color, get_bottom_band_color functions
- Move calc_split_ratio, calc_top_phase, calc_bottom_phase functions
- Move paint_codex_background, paint_codex_blobs functions
- Move paint_band_boundary_diagonal, paint_band_boundary functions
- Move _CODEX_BOUNDARY_ANGLE_DEG constant to module level
- Update graphics.py to import and delegate to codex.background
- Update Claude theme to use codex_bg.blend_colors

The Codex theme now has proper separation with animation phase
calculation, color blending, two-band gradient rendering, and blob
overlay all contained in its own module.

Task: task-028-extract-codex-theme.md

* [FIX] Correct type hints for theme rect parameters

- Change rect parameter type from QWidget to QRect in all theme modules
- Fixes AttributeError: 'QRect' object has no attribute 'rect'
- Update copilot theme to use QRect instead of QWidget
- Update imports to include QRect and remove unused QWidget
- Affects: graphics.py, claude/background.py, codex/background.py, gemini/background.py, copilot/background.py
- All four themes now load and render correctly without errors

* [DOCS] Add task documentation for theme refactoring (tasks 024-031)

Add comprehensive task documentation for the theme extraction and
organization work. Task files track the planning, implementation, and
validation steps for refactoring UI theme code from graphics.py into
dedicated theme modules.

Task files added:
- task-024: Master task and theme structure setup
- task-025: Gemini theme extraction
- task-026: Claude theme extraction
- task-027: Copilot theme extraction
- task-028: Codex theme extraction
- task-029: Graphics cleanup
- task-030: Integration testing
- task-031: Final validation and PR prep

These files provide context for the refactoring work and serve as
implementation guides for future similar work.

* [CHORE] Move completed theme refactor tasks (024-031) to done folder

* [FIX] Ensure copilot sources are initialized in GlassRoot

---------
