+++ .agents/tasks/done/task-031-final-validation-and-pr-prep.md
@@ -1,0 +1,74 @@
+# Task 031: Final Validation and PR Preparation
+
+## Parent Task
+task-024-organize-theme-implementations (master)
+
+## Description
+Perform comprehensive final validation of the theme refactoring work and prepare the branch for PR submission. This includes code quality checks, functionality testing, documentation review, and clean commit history.
+
+## Specific Actions
+
+### 1. Code Quality Validation
+- Run Python linter/type checker on modified files
+- Verify no syntax errors or type hint issues
+- Check for unused imports or variables
+- Ensure code follows project style guidelines (AGENTS.md)
+
+### 2. Functionality Testing
+- Test each theme individually with `uv run main.py --agent [codex|claude|gemini|copilot]`
+- Verify all animations render correctly
+- Check for console errors or warnings
+- Test window resize, minimize, and restore
+- Verify performance metrics are acceptable
+
+### 3. File Organization Review
+- Confirm all theme modules are in correct locations
+- Verify __init__.py files are present and correct
+- Check that no obsolete files remain
+- Ensure imports are clean and follow conventions
+
+### 4. Documentation Check
+- Review commit messages for clarity and format
+- Verify task files are properly organized
+- Ensure no sensitive data in commits
+- Check that AGENTS.md guidelines were followed
+
+### 5. Git History Cleanup
+- Review commit history for logical progression
+- Ensure commit messages follow [TYPE] format
+- Verify no merge conflicts or issues
+- Check branch is up to date with main if needed
+
+### 6. PR Preparation
+- Create summary of changes made
+- List files modified, added, and deleted
+- Document testing performed
+- Note any known issues or follow-up items
+- Add untracked task files to git if they should be included
+
+## Code Location
+- All files in `agents_runner/ui/themes/`
+- Modified files in `agents_runner/ui/`
+- Task files in `.agents/tasks/wip/`
+
+## Technical Context
+- This is the final step before PR submission
+- All previous tasks (024-030) should be complete
+- Branch: midoriaiagents/32e275a4cb
+- 5 commits on branch
+- Major refactoring: extracted themes to dedicated modules
+
+## Dependencies
+- All tasks 024-030 must be complete
+- Code must be tested and working
+- No blocking issues
+
+## Acceptance Criteria
+- All code passes linting/type checking
+- All four themes work correctly
+- No console errors during normal operation
+- Commit history is clean and logical
+- All task files are properly tracked or staged
+- PR summary is complete and accurate
+- Changes align with AGENTS.md guidelines
+- Ready for code review and merge
