+++ agents_runner/ui/themes/__init__.py
@@ -1,0 +1,1 @@
+"""Agent-specific theme implementations for the UI background system."""
