+++ agents_runner/ui/themes/claude/__init__.py
@@ -1,0 +1,1 @@
+"""Background theme implementation for Claude agent."""
