+++ agents_runner/ui/themes/codex/__init__.py
@@ -1,0 +1,1 @@
+"""Background theme implementation for Codex agent."""
