+++ agents_runner/ui/themes/copilot/__init__.py
@@ -1,0 +1,1 @@
+"""Background theme implementation for Copilot agent."""
