+++ agents_runner/ui/themes/gemini/__init__.py
@@ -1,0 +1,1 @@
+"""Background theme implementation for Gemini agent."""
